+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.39
 
+- Add: [YieldFi](https://tradingstrategy.ai/trading-view/vaults/protocols/yieldfi) vyUSD vault on Base and yUSD vault on Ethereum (2026-01-23)
 - Add: New protocol: [Avant](https://www.avantprotocol.com/) - decentralised stablecoin protocol on Avalanche with savUSD staking vault (2026-01-19)
 - Add: New protocol: [Renalta](https://renalta.com/) - yield protocol on Base blockchain with unverified smart contract source code (2026-01-19)
 - Add: New protocol: [infiniFi](https://infinifi.xyz/) - on-chain fractional reserve banking protocol with siUSD liquid staking vault on Ethereum (2026-01-18)